feat: add convertBase64 to utils
