+++ forms/src/utils/convertBase64.ts
@@ -1,0 +1,14 @@
+const getBase64 = (file: File) => {
+  return new Promise<string>((resolve, reject) => {
+    const reader = new FileReader();
+    reader.readAsDataURL(file);
+    reader.onload = function () {
+      resolve(reader.result as string);
+    };
+    reader.onerror = function (error) {
+      reject(error);
+    };
+  });
+};
+
+export { getBase64 };
